FAT-1666 Added RU localization. Restructured project - all activities were replaced into new package named 'activities.with.textview'
