--- app/src/main/java/com/andersen/dogsapp/AbsLayoutParams.java
@@ -1,4 +1,0 @@
-package com.andersen.dogsapp;
-
-public abstract class AbsLayoutParams {
-}
