--- app/src/main/java/com/andersen/dogsapp/TextViewInRootLayoutFactory.java
@@ -1,6 +1,0 @@
-package com.andersen.dogsapp;
-
-public class TextViewInRootLayoutFactory {
-    AbsLayoutParams absLayoutParams = null;
-
-}
