+++ app/src/main/java/com/andersen/dogsapp/activities/with/textview/AbsLayoutParams.java
@@ -1,0 +1,4 @@
+package com.andersen.dogsapp.activities.with.textview;
+
+public abstract class AbsLayoutParams {
+}
