+++ app/src/main/java/com/andersen/dogsapp/activities/with/textview/ActivityTwo.java
@@ -1,7 +1,10 @@
-package com.andersen.dogsapp;
+package com.andersen.dogsapp.activities.with.textview;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
+
+import com.andersen.dogsapp.R;
+
 public class ActivityTwo extends AppCompatActivity {
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
