+++ app/src/main/java/com/andersen/dogsapp/activities/with/textview/Parametrizable.java
@@ -1,4 +1,4 @@
-package com.andersen.dogsapp;
+package com.andersen.dogsapp.activities.with.textview;
 
 public interface Parametrizable<T> {
     Class<T> createLayoutParams(T n, int hor, int ver);
