+++ app/src/main/java/com/andersen/dogsapp/activities/with/textview/TextViewInRootLayoutFactory.java
@@ -1,0 +1,8 @@
+package com.andersen.dogsapp.activities.with.textview;
+
+import com.andersen.dogsapp.activities.with.textview.AbsLayoutParams;
+
+public class TextViewInRootLayoutFactory {
+    AbsLayoutParams absLayoutParams = null;
+
+}
