+++ app/src/main/res/layout/activity_sixth.xml
@@ -5,6 +5,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ActTextViewInConstLayout">
+    tools:context=".activities.with.textview.ActTextViewInConstLayout">
 
 </android.support.constraint.ConstraintLayout>
