+++ app/src/main/res/layout/activity_three.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ActivityThree">
+    tools:context=".activities.with.textview.ActivityThree">
 
     <TextView
         android:id="@+id/textViewDogAge"
